requirements.txt
streamlit
google-generativeai
pyclips
